+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Fix serialization of `ThreeDSecureLookupAdditionalInformation`
+* Return `AuthenticationInsight` when requested on `PaymentMethodNonce.create`
 
 ## 2.96.0
 * Add `ThreeDSecureLookup`